Generalized handling of "services" property
The services is an open set, any software on the web can add their own inbound
and outbound services, so lets handle unknown services a bit better. This adds
ValueOrUnknown helper that tries to either parse the json string into an enum
if possible but still allows unknown string values.